requirements.txt
Flask>=2.0
psycopg2-binary>=2.9 # Usar a versão binária facilita a instalação no Docker